styled & reorganized the components
