+++ src/App.css
@@ -5,7 +5,7 @@
 
 *,
 body {
-  margin: 0;
+  margin: 0 10px;
   padding: 0;
   box-sizing: border-box;
   font-family: "Montserrat", sans-serif;
@@ -16,7 +16,7 @@ body {
 
 .app-logo {
   float: left;
-  position: relative;
+  position: absolute;
   margin-top: 20px;
   width: 20%;
   padding-left: 5%;
@@ -25,8 +25,8 @@ body {
 
 .container {
   max-width: 100%;
-  height: 100vh;
-  display: flex;
+  margin-top:100px;
+  display: grid;
   align-items: center;
   justify-content: center;
 }
