+++ src/components/Live.js
@@ -91,6 +91,7 @@ const Live = () => {
 
   return (
     <div className="container">
+      <h3 className='py-5 font-medium leading-tight text-3xl mt-0 mb-2 text-white-600'>Looking to help?</h3>      
       <button
         className="bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded"
         onClick={() => {
