+++ src/components/Navbar.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+function Navbar() {
+    return (
+        <div className='flex flex-wrap items-center justify-between w-full py-30 mb-24 md:py-20 px-0 text-lg text-gray-700 bg-black'>
+            <img className="app-logo" alt="Logo" src="./Kodelive.png" />          
+        </div>
+    );
+}
+
+export default Navbar;
