+++ src/components/filtertag.js
@@ -1,43 +1,46 @@
-import python from './img/python.png';
-import javascript from './img/javascript.png';
-import sql from './img/sql.png';
-import react from './img/react.png';
-import java from './img/java.png';
+import profile from '../img/profile.png';
 
 //dummy data in the form of objects in the array
 export default [
     {
         id: 1,
-        image: python,
+        image: profile,
         title: 'Python expert', 
         category: 'Python',
         description: 'Lorem Ipsum Lorem Ipsum Lorem Ipsum'
     },
     {
         id: 2,
-        image: javascript,
+        image: profile,
         title: 'Javascript expert', 
         category: 'Javascript',
         description: 'Lorem Ipsum Lorem Ipsum Lorem Ipsum'
     },
     {
         id: 3,
-        image: sql,
+        image: profile,
         title: 'SQL expert', 
         category: 'SQL',
         description: 'Lorem Ipsum Lorem Ipsum Lorem Ipsum'
     },
     {
         id: 4,
-        image: java,
+        image: profile,
         title: 'Java expert', 
         category: 'Java',
         description: 'Lorem Ipsum Lorem Ipsum Lorem Ipsum'
     },
     {
         id: 5,
-        image: react,
+        image: profile,
         title: 'React expert', 
+        category: 'React',
+        description: 'Lorem Ipsum Lorem Ipsum Lorem Ipsum'
+    },
+    {
+        id: 6,
+        image: profile,
+        title: 'React expert #2', 
         category: 'React',
         description: 'Lorem Ipsum Lorem Ipsum Lorem Ipsum'
     }
